requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.7.22
Django==4.2.1
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.5
flake8==6.1.0
flake8-docstrings==1.7.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
packaging==23.2
Pillow==10.0.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
python-gettext==5.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sentry-sdk==1.34.0
snowballstemmer==2.2.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
